requirements_extractor.txt
lxml
numpy~=1.21.1
beautifulsoup4~=4.9.3
graypy~=2.1.0
pydantic~=1.8.2
nltk~=3.6.2
PyYAML~=5.4.1
requests
mongomock~=3.23.0
pymongo~=3.12.0
PyRSMQ
sklearn